Combined the encoder and decoder. Removed the editingString property
