+++ Source/StringTransformer.swift
@@ -1,20 +1,21 @@
 //
-//  Encoder.swift
+//  StringTransformer.swift
 //  SKStringFormatter
 //
-//  Created by Sean on 6/15/17.
+//  Created by Sean on 6/22/17.
 //  Copyright © 2017 Sean Kladek. All rights reserved.
 //
 
 import Foundation
 
-protocol Encoding {
+protocol Transforming {
     func insertFormatStrings(_ inputString: String) -> String
+    func removeFormatStrings(_ inputString: String) -> String
     func trimDisallowedCharacters(_ inputString: String) -> String
     func trimToMaxLength(_ inputString: String) -> String
 }
 
-class Encoder: Encoding {
+class StringTransformer: Transforming {
     let format: StringFormat
 
     init(stringFormat: StringFormat) {
@@ -42,6 +43,31 @@ class Encoder: Encoding {
         return outputString
     }
 
+    func removeFormatStrings(_ inputString: String) -> String {
+        guard var formatStrings = format.formatStrings else {
+            return inputString
+        }
+
+        formatStrings = formatStrings.sorted()
+
+        var outputString = inputString
+        for formatString in formatStrings {
+            let startIndex = outputString.startIndex
+            let endIndex = outputString.endIndex
+            if let lowerBound = outputString.index(startIndex, offsetBy: Int(formatString.startIndex), limitedBy: endIndex) {
+                if let upperBound = outputString.index(lowerBound, offsetBy: formatString.string.characters.count, limitedBy: endIndex) {
+                    let formatRange = lowerBound..<upperBound
+                    let substring = outputString.substring(with: formatRange)
+                    if substring == formatString.string {
+                        outputString = outputString.replacingCharacters(in: formatRange, with: "")
+                    }
+                }
+            }
+        }
+
+        return outputString
+    }
+
     func trimDisallowedCharacters(_ inputString: String) -> String {
         guard let characterSet = format.allowedCharacterSet else {
             return inputString
@@ -62,7 +88,7 @@ class Encoder: Encoding {
         let stringStartIndex = inputString.startIndex
         let endIndex = inputString.index(stringStartIndex, offsetBy: format.maxLength)
         let outputString = inputString.substring(to: endIndex)
-
+        
         return outputString
     }
 }
